% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{ag_cut}
\alias{ag_cut}
\title{A simple method to set age groups by intervals}
\usage{
ag_cut(x, low_breaks = 19, high_breaks = 70, interval = 10, to_char = "-")
}
\arguments{
\item{x}{Any vector.}

\item{low_breaks}{Determine the first age group.}

\item{high_breaks}{Determine the last age group.}

\item{interval}{Determine the intervals of age groups.}

\item{to_char}{Set the connector of age groups (start and end) excluding the first and last.}
}
\description{
A simple method to set age groups by intervals
}
\examples{
age <- c(1:100, 1:100)
ag_cut(age)
}
