% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{age_cut}
\alias{age_cut}
\title{A simple method to set age groups by intervals}
\usage{
age_cut(x, low.breaks = 19, high.breaks = 70, interval = 10, to_char = "-")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
A simple method to set age groups by intervals
}
\examples{
age <- 1:100
age_cut(age)
}
