% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{easy_clust}
\alias{easy_clust}
\title{One-click clustering from feature selection to
visulization and correlation analysis}
\usage{
easy_clust(x, meta = NULL)
}
\description{
One-click clustering from feature selection to
visulization and correlation analysis
}
