% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{gsea_senmayo}
\alias{gsea_senmayo}
\title{GSEA for SenMayo geneset}
\usage{
gsea_senmayo(geneList, type = "Geneid", ...)
}
\value{
list
}
\description{
GSEA for SenMayo geneset
}
\examples{
data(geneList, package="DOSE")
gene <- geneList[abs(geneList) > 2]
res <- gsea_senmayo(gene,'Geneid',pvalueCutoff=1, minGSSize =1) # for all result (no-sign result include)
}
