% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{gsva_senmayo}
\alias{gsva_senmayo}
\title{GSVA (orssGSEA) for SenMayo geneset for each cell or sample
obj sample x gene}
\usage{
gsva_senmayo(obj, ...)
}
\value{
list
}
\description{
GSVA (orssGSEA) for SenMayo geneset for each cell or sample
obj sample x gene
}
\examples{
tpm_sample <- data(tpm_sample)
gsva_senmayo(t(tpm_sample),groups = 10, cores = 2)
}
