% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{SenMayo_GSEA}
\alias{SenMayo_GSEA}
\title{GSEA for SenMayo geneset}
\usage{
SenMayo_GSEA(geneList, type = "Geneid", ...)
}
\value{
list
}
\description{
GSEA for SenMayo geneset
}
\examples{
data(geneList, package="DOSE")
gene <- geneList[abs(geneList) > 2]
res <- SenMayo_GSEA(gene,'Geneid',pvalueCutoff=1, minGSSize =1) # for all result (no-sign result include)
}
