% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{SenMayo_GSVA}
\alias{SenMayo_GSVA}
\title{GSVA (orssGSEA) for SenMayo geneset for each cell or sample
obj sample x gene}
\usage{
SenMayo_GSVA(obj, ...)
}
\value{
list
}
\description{
GSVA (orssGSEA) for SenMayo geneset for each cell or sample
obj sample x gene
}
\examples{
tpm_sample <- data(tpm_sample)
SenMayo_GSVA(t(tpm_sample),groups = 10, cores = 2)
}
