% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SenMayoName}
\alias{SenMayoName}
\title{SenMayoName aging gene set gene name table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125 rows and 2 columns.
}
\source{
A new gene set identifies senescent cells and predicts senescence-associated pathways across tissues
}
\usage{
SenMayoName
}
\description{
SenMayoName aging gene set gene name table
}
\keyword{datasets}
