% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SenMayoSet}
\alias{SenMayoSet}
\title{SenMayoSet raw dataset}
\format{
An object of class \code{list} of length 1.
}
\source{
A new gene set identifies senescent cells and predicts senescence-associated pathways across tissues
}
\usage{
SenMayoSet
}
\description{
SenMayoSet raw dataset
}
\keyword{datasets}
