% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{vis_bar}
\alias{vis_bar}
\title{Draw bar plot by N of age groups.}
\usage{
vis_bar(
  x,
  age_cut_col = "age_cut",
  age_col = "age",
  facet_var = NULL,
  fill = NULL,
  convert_to_percent = FALSE,
  show_zero = FALSE,
  palt = NULL,
  bar_val = NULL,
  add_param = list(title = "All samples", title_median = TRUE, title_iqr = TRUE, digits =
    0, add_line = "median", linecol = "gray", linetype = "dashed"),
  ...
)
}
\description{
Draw bar plot by N of age groups.
}
\examples{
age <- data.frame(age=rgamma(10000, shape = 50), groups = "Age-50")
age$age_cut <- ag_cut(age$age, low_breaks = 29)
vis_bar(age, age_cut_col = "age_cut")
age2 <- data.frame(age=rgamma(10000, shape = 70), groups = "Age-70")
age2$age_cut <- ag_cut(age2$age, low_breaks = 29)
age <- rbind(age, age2)
vis_bar(age, age_cut_col = "age_cut", facet_var = "groups")
}
