% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{vis_bar_logic}
\alias{vis_bar_logic}
\title{Draw bar percent plot with logistic regression}
\usage{
vis_bar_logic(
  x,
  factor_var,
  age_cut = "age_cut",
  facet_var = NULL,
  title_prefix = "",
  xlab = "Age groups",
  ylab = "Percent",
  digits = 3,
  palt = ag_col(theme = "pal_npg"),
  ...
)
}
\description{
Draw bar percent plot with logistic regression
}
\examples{
age <- data.frame(age=rgamma(10000, shape = 50), groups = "Age-50")
age$age_cut <- ag_cut(age$age)
age$os <- c(rgamma(2000, shape = 160), rgamma(2000, shape = 320),
  rgamma(4000, shape = 560), rgamma(2000, shape = 800))
age$os_status <- round(rgamma(10000, shape = 720) \%\% 2)
age$os_status[age$os_status != 0] <- 1
age$facet_var <- round(rgamma(10000, shape = 720) \%\% 2)
age$facet_var[age$facet_var != 0] <- 1
vis_bar_logic(age, "os_status")
}
