% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{vis_cox}
\alias{vis_cox}
\title{Ezcox for multivariate analysis}
\usage{
vis_cox(x, ret = c("forest_plot"), covariates = "age_cut", ...)
}
\description{
Ezcox for multivariate analysis
}
\examples{
age <- data.frame(age=rgamma(10000, shape = 50), groups = "Age-50")
age$age_cut <- ag_cut(age$age)
age$os <- c(rgamma(2000, shape = 160), rgamma(2000, shape = 320),
  rgamma(4000, shape = 560), rgamma(2000, shape = 800))
age$os_status <- round(rgamma(10000, shape = 720) \%\% 2)
age$os_status[age$os_status != 0] <- 1
age$gender <- round(rgamma(10000, shape = 720) \%\% 2)
age$gender[age$gender != 0] <- 1
vis_cox(age, time = "os", status = "os_status", covariates = "age_cut", controls = c("gender"))
vis_cox(age, ret = c("forest_plot", "model"), time = "os",
       status = "os_status", covariates = "age_cut", controls = c("gender"))
}
