% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{vis_mut_line}
\alias{vis_mut_line}
\title{Percent line plots}
\usage{
vis_mut_line(
  x,
  maf,
  age = "age",
  fill = "fusion_genes",
  sample_id = "Tumor_Sample_Barcode",
  gene = "Hugo_Symbol",
  type = "Variant_Classification"
)
}
\description{
Percent line plots
}
\examples{
maf <- read.csv(system.file("extdata", "aml_maf.txt", package = "agmin"), sep = "\t")
x <- read.csv(system.file("extdata", "aml_samples.txt", package = "agmin"), sep = "\t")
}
