% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{vis_mutc_bar}
\alias{vis_mutc_bar}
\title{Show bar plots of mutation data}
\usage{
vis_mutc_bar(
  maf,
  palt = NULL,
  cutoff = 3,
  sample_id = "Tumor_Sample_Barcode",
  gene = "Hugo_Symbol",
  type = "Variant_Classification"
)
}
\description{
Show bar plots of mutation data
}
\examples{
maf <- read.csv(system.file("extdata", "aml_maf.txt", package = "agmin"), sep = "\t")
print(vis_mutc_bar(maf, gene = "Hugo_Symbol", type = "Variant_Classification"))
}
