% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{vis_mutc_boxplot}
\alias{vis_mutc_boxplot}
\title{Show boxplot plots: mutation counts and groups}
\usage{
vis_mutc_boxplot(
  x,
  maf,
  age = "age",
  facet_var = "fusion_genes",
  fill = "fusion_genes",
  sample_id = "Tumor_Sample_Barcode",
  gene = "Hugo_Symbol",
  type = "Variant_Classification"
)
}
\description{
Show boxplot plots: mutation counts and groups
}
