% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{vis_mutc_percent}
\alias{vis_mutc_percent}
\title{Percent barplot}
\usage{
vis_mutc_percent(
  x,
  maf,
  age = "age",
  facet_var = "fusion_genes",
  fill = "fusion_genes",
  palt = NULL,
  sample_id = "Tumor_Sample_Barcode",
  gene = "Hugo_Symbol",
  type = "Variant_Classification",
  logist_mutc = 4
)
}
\description{
Percent barplot
}
