% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{vis_mutc_scatter}
\alias{vis_mutc_scatter}
\title{Show scatter plots: mutation counts and age}
\usage{
vis_mutc_scatter(
  x,
  maf,
  age = "age",
  facet_var = "fusion_genes",
  fill = "fusion_genes",
  sample_id = "Tumor_Sample_Barcode",
  gene = "Hugo_Symbol",
  type = "Variant_Classification"
)
}
\description{
Show scatter plots: mutation counts and age
}
