% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{vis_surv}
\alias{vis_surv}
\title{Draw survival curve by age groups}
\usage{
vis_surv(
  x,
  group = "age_cut",
  time = "os",
  status = "os_status",
  title = "",
  xlab = "Time in days",
  leg_labs = NULL,
  xlim = NULL,
  plt = NULL,
  brk = 180,
  rtb = TRUE,
  pval = TRUE,
  conf_int = TRUE,
  conf_alpha = 0.1,
  nc_plot = FALSE,
  nc_plot_h = 0.3,
  rtb_y_col = TRUE,
  rtb_y = FALSE,
  rtb_h = 0.25,
  ...
)
}
\description{
Draw survival curve by age groups
}
\examples{
age <- data.frame(age=rgamma(10000, shape = 50), groups = "Age-50")
age$age_cut <- ag_cut(age$age)
age$os <- c(rgamma(2000, shape = 160), rgamma(2000, shape = 320),
  rgamma(4000, shape = 560), rgamma(2000, shape = 800))
age$os_status <- round(rgamma(10000, shape = 720) \%\% 2)
age$os_status[age$os_status != 0] <- 1
vis_surv(age, plt = ag_col()[-c(1,2)])
}
